+++ okitweb/static/okit/view/designer/js/artefacts/internet_gateway.js
@@ -24,17 +24,8 @@ class InternetGatewayView extends OkitDesignerArtefactView {
     /*
     ** Property Sheet Load function
      */
-    loadProperties() {
-        let okitJson = this.getOkitJson();
-        let me = this;
-        $(jqId(PROPERTIES_PANEL)).load("propertysheets/internet_gateway.html", () => {loadPropertiesSheet(me.artefact);});
-    }
-
-    /*
-    ** Load and display Value Proposition
-     */
-    loadValueProposition() {
-        $(jqId(VALUE_PROPOSITION_PANEL)).load("valueproposition/internet_gateway.html");
+    newPropertiesSheet() {
+        this.properties_sheet = new InternetGatewayProperties(this.artefact)
     }
 
     /*